requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
django-filter
Faker
behave
behave-django
selenium
gunicorn
pandas
requests
drf-yasg
lxml
django-crispy-forms